+++ src/core/watchers/MidaMarketWatcherParameters.ts
@@ -29,4 +29,10 @@ import { MidaMarketWatcher, } from "#watchers/MidaMarketWatcher";
  */
 export type MidaMarketWatcherParameters = {
     tradingAccount: MidaTradingAccount;
+    /**
+     * Indicates if the market watcher should check manually for candlesticks being closed
+     * This is useful when a trading platform doesn't provide events for candlesticks being closed
+     * This leverages the getSymbolPeriods() API
+     */
+    useBuiltinPeriodCloseDetector?: boolean;
 };